Output a css file instead of a less file
Instead of generating a Less file that will need to be compiled anyway
and doesn’t have the correct behavior when trying to “extend” the
Bootstrap styles, generate a CSS file. Use a .import.less on the output
file so it isn’t automatically processed.

Also trigger the build plugin whenever the Bootstrap variables are
changed